Fixed keymaps, broken link, encoder handling
